Move StickyConfirmButtons to its own component
